Improve hardware lock elision
We must use memory_order_acquire in predicates like is_locked(),
because they may be consulted in lock elision.

In the test program, we must define a special transactional_assert()
that will commit the memory transaction before aborting the execution.
Invoking normal assert() would invoke a system call, which would
trivially result in an abort of the memory transaction.