Common s390x and POWER8+ HTM implementation
Notes on the non-line implementation.

Both ppc64 and s390x will only define the low level
__builtin_t{begin,end,abort} macros when -mhtm is
specified as a compile flag. An attribute target("htm")
isn't sufficient, though it has been included in
the hope this gcc limitation is ultimately fixed.

s390x has a limitation that its tbegin cannot be
inline. This is documented in their xhtmintrin.h
header file (as a FIXME).

SIGILL implementation model of openssl.