Remove some asm work-arounds
Starting with GCC 7 and clang 15, single-bit std::atomic::fetch_or()
actually works, except for the most significant bit, which was
fixed in GCC 13 and as of now, not in clang yet:
https://github.com/llvm/llvm-project/issues/37322

The Microsoft compiler still prefers to emit loops around LOCK CMPXCHG,
so we will keep using another work-around on that platform.

It should be noted that GCC 7 is the oldest compiler available in
currently supported GNU/Linux distributions.  FreeBSD 14 uses clang 16.